+++ src/percipio_camera/src/gige_2_1.cpp
@@ -257,13 +257,13 @@ void GigE_2_1::device_load_parameters()
                 }
         
                 case Float: {
-                    double val = atoi(feat_val.c_str());
+                    double val = atof(feat_val.c_str());
                     status = TYFloatSetValue(hDevice, feature.c_str(), val);
                     break;
                 }
 
                 case Boolean: {
-                    bool val = static_cast<bool>(atof(feat_val.c_str()));
+                    bool val = static_cast<bool>(atoi(feat_val.c_str()));
                     status = TYBooleanSetValue(hDevice, feature.c_str(), val);
                     break;
                 }
